requirements.txt
bottle
bcrypt
bottle-beaker
pylint